# specs/caja-openapi.yaml
openapi: 3.0.3
info:
  title: eParking - Caja & Turnos API
  version: '1.0'
servers:
  - url: http://localhost:3010/api
paths:
...
  /tickets/{id}:
    get:
      summary: Get ticket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Ticket:
      type: object
      properties:
        _id: { type: string }
        shiftId: { type: string }
        storeId: { type: string }
        createdBy: { type: string }
        createdAt: { type: string }
        status: { type: string }
        total: { type: integer }
        items: { type: array }
        paymentMethod: { type: string }
        payments: { type: array }
        notes: { type: string }
...